+++ scalaz-effect/src/main/scala/loggerf/scalaz/LoggerEitherT.scala
@@ -1,9 +1,11 @@
-package loggerf
+package loggerf.scalaz
 
-import cats._
-import cats.data.EitherT
-import cats.implicits._
-import effectie.cats.EffectConstructor
+import scalaz._
+import Scalaz._
+
+import effectie.scalaz.EffectConstructor
+
+import loggerf.Logger
 
 trait LoggerEitherT[F[_]] {
 
@@ -19,11 +21,11 @@ trait LoggerEitherT[F[_]] {
   , b2String: B => String
   ): EitherT[F, A, B] =
     EitherT(
-      MF0.flatMap(efab.value) {
-        case Left(a) =>
-          EF0.effectOf(logger0.debug(a2String(a))) *> EF0.effectOf(a.asLeft[B])
-        case Right(b) =>
-          EF0.effectOf(logger0.debug(b2String(b))) *> EF0.effectOf(b.asRight[A])
+      MF0.bind(efab.run) {
+        case -\/(a) =>
+          EF0.effectOf(logger0.debug(a2String(a))) *> EF0.effectOf(a.left[B])
+        case \/-(b) =>
+          EF0.effectOf(logger0.debug(b2String(b))) *> EF0.effectOf(b.right[A])
       }
     )
 
@@ -34,11 +36,11 @@ trait LoggerEitherT[F[_]] {
   , b2String: B => String
   ): EitherT[F, A, B] =
     EitherT(
-      MF0.flatMap(efab.value) {
-        case Left(a) =>
-          EF0.effectOf(logger0.info(a2String(a))) *> EF0.effectOf(a.asLeft[B])
-        case Right(b) =>
-          EF0.effectOf(logger0.info(b2String(b))) *> EF0.effectOf(b.asRight[A])
+      MF0.bind(efab.run) {
+        case -\/(a) =>
+          EF0.effectOf(logger0.info(a2String(a))) *> EF0.effectOf(a.left[B])
+        case \/-(b) =>
+          EF0.effectOf(logger0.info(b2String(b))) *> EF0.effectOf(b.right[A])
       }
     )
 
@@ -49,11 +51,11 @@ trait LoggerEitherT[F[_]] {
   , b2String: B => String
   ): EitherT[F, A, B] =
     EitherT(
-      MF0.flatMap(efab.value) {
-        case Left(a) =>
-          EF0.effectOf(logger0.warn(a2String(a))) *> EF0.effectOf(a.asLeft[B])
-        case Right(b) =>
-          EF0.effectOf(logger0.warn(b2String(b))) *> EF0.effectOf(b.asRight[A])
+      MF0.bind(efab.run) {
+        case -\/(a) =>
+          EF0.effectOf(logger0.warn(a2String(a))) *> EF0.effectOf(a.left[B])
+        case \/-(b) =>
+          EF0.effectOf(logger0.warn(b2String(b))) *> EF0.effectOf(b.right[A])
       }
     )
 
@@ -64,11 +66,11 @@ trait LoggerEitherT[F[_]] {
   , b2String: B => String
   ): EitherT[F, A, B] =
     EitherT(
-      MF0.flatMap(efab.value) {
-        case Left(a) =>
-          EF0.effectOf(logger0.error(a2String(a))) *> EF0.effectOf(a.asLeft[B])
-        case Right(b) =>
-          EF0.effectOf(logger0.error(b2String(b))) *> EF0.effectOf(b.asRight[A])
+      MF0.bind(efab.run) {
+        case -\/(a) =>
+          EF0.effectOf(logger0.error(a2String(a))) *> EF0.effectOf(a.left[B])
+        case \/-(b) =>
+          EF0.effectOf(logger0.error(b2String(b))) *> EF0.effectOf(b.right[A])
       }
     )
 }
