+++ scalaz-effect/src/main/scala/loggerf/scalaz/LoggerOptionT.scala
@@ -1,63 +1,94 @@
-package loggerf
+package loggerf.scalaz
 
-import cats._
-import cats.implicits._
+import scalaz._
+import Scalaz._
 
-import effectie.cats.EffectConstructor
+import effectie.scalaz.EffectConstructor
 
-trait LoggerOption[F[_]] {
+import loggerf.Logger
+
+trait LoggerOptionT[F[_]] {
 
   implicit val EF0: EffectConstructor[F]
   implicit val MF0: Monad[F]
 
   implicit val logger0: Logger
 
-  def debugOption[A](fa: F[Option[A]])(ifEmpty: => String, a2String: A => String): F[Option[A]] =
-    MF0.flatMap(fa) {
+  def debugOptionT[A](
+    ofa: OptionT[F, A]
+  )(
+    ifEmpty: => String
+  , a2String: A => String
+  ): OptionT[F, A] =
+    OptionT(
+      MF0.bind(ofa.run) {
       case Some(a) =>
         EF0.effectOf(logger0.debug(a2String(a))) *> EF0.effectOf(a.some)
       case None =>
         EF0.effectOf(logger0.debug(ifEmpty)) *> EF0.effectOf(none[A])
-    }
+      }
+    )
 
-  def infoOption[A](fa: F[Option[A]])(ifEmpty: => String, a2String: A => String): F[Option[A]] =
-    MF0.flatMap(fa) {
+  def infoOptionT[A](
+    ofa: OptionT[F, A]
+  )(
+    ifEmpty: => String
+  , a2String: A => String
+  ): OptionT[F, A] =
+    OptionT(
+      MF0.bind(ofa.run) {
       case Some(a) =>
         EF0.effectOf(logger0.info(a2String(a))) *> EF0.effectOf(a.some)
       case None =>
         EF0.effectOf(logger0.info(ifEmpty)) *> EF0.effectOf(none[A])
-    }
+      }
+    )
 
-  def warnOption[A](fa: F[Option[A]])(ifEmpty: => String, a2String: A => String): F[Option[A]] =
-    MF0.flatMap(fa) {
+  def warnOptionT[A](
+    ofa: OptionT[F, A]
+  )(
+    ifEmpty: => String
+  , a2String: A => String
+  ): OptionT[F, A] =
+    OptionT(
+      MF0.bind(ofa.run) {
       case Some(a) =>
         EF0.effectOf(logger0.warn(a2String(a))) *> EF0.effectOf(a.some)
       case None =>
         EF0.effectOf(logger0.warn(ifEmpty)) *> EF0.effectOf(none[A])
-    }
+      }
+    )
 
-  def errorOption[A](fa: F[Option[A]])(ifEmpty: => String, a2String: A => String): F[Option[A]] =
-    MF0.flatMap(fa) {
+  def errorOptionT[A](
+    ofa: OptionT[F, A]
+  )(
+    ifEmpty: => String
+  , a2String: A => String
+  ): OptionT[F, A] =
+    OptionT(
+      MF0.bind(ofa.run) {
       case Some(a) =>
         EF0.effectOf(logger0.error(a2String(a))) *> EF0.effectOf(a.some)
       case None =>
         EF0.effectOf(logger0.error(ifEmpty)) *> EF0.effectOf(none[A])
-    }
+      }
+    )
+
 }
 
-object LoggerOption {
-  def apply[F[_] : LoggerOption]: LoggerOption[F] = implicitly[LoggerOption[F]]
+object LoggerOptionT {
+  def apply[F[_] : LoggerOptionT]: LoggerOptionT[F] = implicitly[LoggerOptionT[F]]
 
   @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
-  implicit def loggerOption[F[_]](
+  implicit def loggerOptionT[F[_]](
     implicit EF: EffectConstructor[F], MF: Monad[F], logger: Logger
-  ): LoggerOption[F] = new LoggerOptionF[F]
+  ): LoggerOptionT[F] = new LoggerOptionTF[F]
 
-  final class LoggerOptionF[F[_]](
+  final class LoggerOptionTF[F[_]](
     @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
     implicit override val EF0: EffectConstructor[F]
   , override val MF0: Monad[F]
   , override val logger0: Logger
-  ) extends LoggerOption[F]
+  ) extends LoggerOptionT[F]
 
 }
